requirements.txt
pyrofork
fastapi
uvicorn
pymongo
python-dotenv
TgCrypto